Several changes based on PR feedback
Move upstream matchers into spec/support/matchers.rb instead of
libraries/matchers.

Included create_server parameters and fixed conditional.

Removed duplicate server_spec file for default recipe.

Began adding serverspec tests (but not much progress yet.)
