--- test/integration/default/serverspec/server_spec.rb
@@ -1,3 +1,0 @@
-require 'serverspec'
-
-set :backend, :exec
